Move setattr() and unlink() to metadata.
